Port to hyper 0.13 and async/await
In this commit, we replace the outdated `hyper-native-tls` dependency
with `hyper-tls` (which also utilizes `native-tls` under the
hood).

We also add the `headers` crate for reintroducing the typed HTTP header
types that were removed from `hyper` proper in 2019, as suggested by:

https://github.com/hyperium/hyper/issues/1950

The signature for `Client::respond_to_webhook()` has changed to match
upstream changes to the `hyper` and `http` crates.